refactor: 일정 조회 후, 명세된 정보를 바탕으로 4가지 경우에 맞게 렌더링하여 일정 데이터 생성하도록 구조 개선
1. 기간 X + 반복 X → 하루만 표시
2. 기간 X + 반복 O → 모든 날짜 중 반복 요일에 포함된 날만
3. 기간 O + 반복 X → 기간 내 모든 날짜
4. 기간 O + 반복 O → 기간 내의 반복 요일에만 표시
