+++ Projects/TDPresentation/Sources/AppFlow/MainFlow/Calendar/ToduckCalendarCoordinator.swift
@@ -18,12 +18,10 @@ final class ToduckCalendarCoordinator: Coordinator {
 
     func start() {
         let fetchScheduleListUseCase = injector.resolve(FetchScheduleListUseCase.self)
-        let buildMonthScheduleDictUseCase = injector.resolve(BuildMonthScheduleDictUseCase.self)
         let finishScheduleUseCase = DIContainer.shared.resolve(FinishScheduleUseCase.self)
         let finishRoutineUseCase = DIContainer.shared.resolve(FinishRoutineUseCase.self)
         let toduckCalendarViewModel = ToduckCalendarViewModel(
             fetchScheduleListUseCase: fetchScheduleListUseCase,
-            buildMonthScheduleDictUseCase: buildMonthScheduleDictUseCase,
             finishScheduleUseCase: finishScheduleUseCase,
             finishRoutineUseCase: finishRoutineUseCase
         )
