+++ Projects/TDPresentation/Sources/AppFlow/MainFlow/Home/Toduck/ToduckViewModel.swift
@@ -61,40 +61,40 @@ final class ToduckViewModel: BaseViewModel {
     }
     
     private func fetchScheduleList() async {
-        do {
-            let todayFormat = Date().convertToString(formatType: .yearMonthDay)
-            let fetchedTodaySchedules = try await fetchScheduleListUseCase.execute(
-                startDate: todayFormat,
-                endDate: todayFormat
-            )
-            isAllDays = shouldMarkAllDayUseCase.execute(with: fetchedTodaySchedules)
-            todaySchedules = fetchedTodaySchedules.sorted { Date.timeSortKey($0.time) < Date.timeSortKey($1.time) }
-
-            // 오늘 날짜에 완료 안 한 일정만 필터링
-            uncompletedSchedules = fetchedTodaySchedules
-                .filter { schedule in
-                    guard let records = schedule.scheduleRecords, !records.isEmpty else {
-                        return true // 오늘 기록이 없으면 완료 안한 상태
-                    }
-                    
-                    if let todayRecord = records.first(where: { $0.recordDate == todayFormat }) {
-                        return !todayRecord.isComplete // 기록이 있다면, 완료안된 것만 포함
-                    } else {
-                        return true // 오늘 기록이 없으면 완료 안한 상태
-                    }
-                }
-                .sorted {
-                    Date.timeSortKey($0.time) < Date.timeSortKey($1.time)
-                }
-
-            if fetchedTodaySchedules.isEmpty {
-                output.send(.fetchedEmptyScheduleList)
-            } else {
-                output.send(.fetchedScheduleList)
-            }
-        } catch {
-            output.send(.failure(error: "일정을 불러오는데 실패했습니다."))
-        }
+//        do {
+//            let todayFormat = Date().convertToString(formatType: .yearMonthDay)
+//            let fetchedTodaySchedules = try await fetchScheduleListUseCase.execute(
+//                startDate: todayFormat,
+//                endDate: todayFormat
+//            )
+//            isAllDays = shouldMarkAllDayUseCase.execute(with: fetchedTodaySchedules)
+//            todaySchedules = fetchedTodaySchedules.sorted { Date.timeSortKey($0.time) < Date.timeSortKey($1.time) }
+//
+//            // 오늘 날짜에 완료 안 한 일정만 필터링
+//            uncompletedSchedules = fetchedTodaySchedules
+//                .filter { schedule in
+//                    guard let records = schedule.scheduleRecords, !records.isEmpty else {
+//                        return true // 오늘 기록이 없으면 완료 안한 상태
+//                    }
+//                    
+//                    if let todayRecord = records.first(where: { $0.recordDate == todayFormat }) {
+//                        return !todayRecord.isComplete // 기록이 있다면, 완료안된 것만 포함
+//                    } else {
+//                        return true // 오늘 기록이 없으면 완료 안한 상태
+//                    }
+//                }
+//                .sorted {
+//                    Date.timeSortKey($0.time) < Date.timeSortKey($1.time)
+//                }
+//
+//            if fetchedTodaySchedules.isEmpty {
+//                output.send(.fetchedEmptyScheduleList)
+//            } else {
+//                output.send(.fetchedScheduleList)
+//            }
+//        } catch {
+//            output.send(.failure(error: "일정을 불러오는데 실패했습니다."))
+//        }
     }
     
     func setSegment(_ type: ScheduleSegmentType) {
